# src/main/resources/openapi/credit-service.yaml
openapi: 3.0.3
info:
  title: Credit Service API
  version: 1.0.0
  description: API for managing credits (personal and enterprise)

servers:
  - url: http://localhost:8080/api/v1

paths:
  /credits:
    get:
      summary: Get all credits
      operationId: getAllCredits
      parameters:
        - name: customerId
          in: query
          required: false
          schema:
            type: string
          description: Filter credits by customer ID
          example: "68a0cdde7d68c910da0bc239"
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
          description: Filter credit by active status (default = true)
          example: true
      responses:
        '200':
          description: List of customers credits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreditResponse'
    post:
      summary: Create a new credit
      operationId: createCredit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCreateRequest'
      responses:
        '201':
          description: Credit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '400':
          description: Bad request - Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                personalCreditExists:
                  summary: Personal customer already has a credit
                  value:
                    code: "PERSON_ALREADY_HAS_CREDIT"
                    message: "Personal customers can only have one active credit"
                    timestamp: "2025-08-24T10:15:30Z"
                customerNotFound:
                  summary: Customer not found
                  value:
                    code: "CUSTOMER_NOT_FOUND"
                    message: "Customer with provided ID does not exist"
                    timestamp: "2025-08-24T10:15:30Z"

  /credits/{id}:
    get:
      summary: Get a credit by ID
      operationId: getCreditById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a credit by ID
      operationId: updateCredit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditUpdateRequest'
      responses:
        '200':
          description: Credit updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request - Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a credit by ID
      operationId: deleteCredit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Credit deleted
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /credits/{id}/deactivate:
    patch:
      summary: Soft delete (deactivate) a credit by ID
      description: Marks the credit as inactive without physically removing it
      operationId: deactivateCredit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit deactivated (isActive set to false)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /credits/{id}/activate:
    patch:
      summary: Reactivate a credit by ID
      description: Marks the credit as active again
      operationId: activateCredit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit activated (isActive set to true)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /credits/{creditNumber}/process-payment:
    post:
      summary: Process payment validation and update balances
      description: Validates payment and updates credit balances if successful
      operationId: processCreditPayment
      parameters:
        - name: creditNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentProcessRequest'
      responses:
        '200':
          description: Payment processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentProcessResponse'
        '400':
          description: Payment validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentProcessResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /credits/{creditNumber}/balance:
    get:
      summary: Get credit balance information
      operationId: getCreditBalance
      parameters:
        - name: creditNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit balance information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditBalanceResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    CreditCreateRequest:
      type: object
      properties:
        customerId:
          type: string
          example: "68a0cdde7d68c910da0bc239"
        creditLimit:
          type: number
          format: double
          example: 10000.00
        availableCredit:
          type: number
          format: double
          example: 7000
      required:
        - customerId
        - creditLimit

    CreditUpdateRequest:
      type: object
      properties:
        creditLimit:
          type: number
          format: double
          example: 12000.00
        availableCredit:
          type: number
          format: double
          example: 8000
        isActive:
          type: boolean
          example: true
      required:
        - creditLimit
        - availableCredit

    PaymentProcessRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
          description: Payment amount
          example: 500.00

    PaymentProcessResponse:
      type: object
      required:
        - success
        - creditId
        - requestedAmount
        - processedAt
      properties:
        success:
          type: boolean
          description: Whether the payment was successful
          example: true
        creditId:
          type: string
          description: Credit ID
          example: "64f3cbb5f8a1e53d2f7c34e1"
        requestedAmount:
          type: number
          format: double
          description: Requested payment amount
          example: 500.00
        actualPaymentAmount:
          type: number
          format: double
          description: Actual amount paid (only if success = true)
          example: 500.00
        availableCreditAfter:
          type: number
          format: double
          description: Available credit after payment (only if success = true)
          example: 9500.00
        #currentBalanceAfter:
          #type: number
          #format: double
          #description: Current balance after payment (only if success = true)
          #example: 500.00
        errorCode:
          type: string
          enum: [ CREDIT_INACTIVE, INVALID_AMOUNT, OVERPAYMENT_NOT_ALLOWED, ZERO_CURRENT_BALANCE ]
          description: Error code (only if success = false)
          example: "OVERPAYMENT_NOT_ALLOWED"
        errorMessage:
          type: string
          description: Error message (only if success = false)
          example: "Payment amount exceeds current balance"
        processedAt:
          type: string
          format: date-time
          description: Processing timestamp
          example: "2025-08-30T15:30:00Z"

    CreditResponse:
      type: object
      properties:
        id:
          type: string
          example: "64f3cbb5f8a1e53d2f7c34e1"
        creditNumber:
          type: string
          example: "PER-0001"
        customerId:
          type: string
          example: "68a0cdde7d68c910da0bc239"
        type:
          type: string
          enum: ["PERSONAL", "ENTERPRISE"]
          example: "PERSONAL"
        creditLimit:
          type: number
          format: double
          example: 10000.00
        availableCredit:
          type: number
          format: double
          example: 7000
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: 2025-08-18T10:15:30Z
        updatedAt:
          type: string
          format: date-time
          example: 2025-08-18T10:15:30Z

    CreditBalanceResponse:
      type: object
      required:
        - creditId
        - creditNumber
        - creditLimit
        - availableCredit
        #- currentBalance
        - utilizationPercentage
      properties:
        creditId:
          type: string
          description: Credit ID
          example: "64f3cbb5f8a1e53d2f7c34e1"
        creditNumber:
          type: string
          description: Credit number
          example: "1234567891234567"
        creditLimit:
          type: number
          format: double
          description: Total credit limit
          example: 10000.00
        availableCredit:
          type: number
          format: double
          description: Available credit
          example: 6000.00
        #currentBalance:
          #type: number
          #format: double
          #description: Current balance (debt)
          #example: 4000.00
        utilizationPercentage:
          type: number
          format: double
          description: Credit utilization percentage
          example: 40.00
        isActive:
          type: boolean
          description: Credit active status
          example: true

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "PERSON_ALREADY_HAS_CREDIT"
        message:
          type: string
          example: "Personal customers can only have one active credit"
        timestamp:
          type: string
          format: date-time
          example: "2025-08-24T10:15:30Z"
      required:
        - code
        - message
        - timestamp